Created SuperClass VehiculoGente as a subclass of Vehiculo. This required a signature change in Supermethod getPrecioAlquiler, it now returns a double instead of a float.

+++ src/Vehiculo.java
@@ -1,7 +1,7 @@
 public class Vehiculo {
     private String matricula;
 
-    public float getPrecioAlquiler(int dias){
+    public double getPrecioAlquiler(int dias){
         return 50*dias;
     }
